Apply suggestions from code review
Hey, many thanks for the review. I was not shure to bother you with this but as I see now, you are watching. If ok, than I will add you to my review requests. Note that atm a review is not required for merging easing to implement small changes quickly, PR's are mandatory for knowing what has happened. For bigger changes we do standard reviewing...

Co-authored-by: Phil Davis <phil@jankaritech.com>