Update upgrading_4.0.0_5.0.0.adoc
"check if you are affected" was ambiguous. In my understanding, I am of course affected, as I run ocis 4 and want ocis 5. "the upgrade affects me". But this is not the point. 

The important warning we have to have in the introduction, is that one should read about all the possibly breaking changes before starting with step 1. 

I am reprhasing the wordning, in the assumption that the latter is meant.